CS4 compilation error
git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@5863 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ notforrelease_cancer/src/mesh/VertexMesh.cpp
@@ -1441,7 +1441,7 @@ void VertexMesh<ELEMENT_DIM, SPACE_DIM>::PerformT2Swap(VertexElement<ELEMENT_DIM
      c_vector<double, SPACE_DIM>& new_node_location = pElement->GetNode(0)->rGetModifiableLocation();    
      new_node_location = GetCentroidOfElement(pElement->GetIndex());
      
-     c_vector<double, 3> neighbouring_elem_nums;
+     c_vector<unsigned, 3> neighbouring_elem_nums;
      
      for (unsigned i=0; i<3; i++)
      {
